+++ src/units/heating-value.ts
@@ -2,6 +2,8 @@ import * as UnitDivide from "../unit-divide";
 import { KiloWattHour } from "./energy2";
 import { CubicMeter } from "./volume";
 
+export type HeatingValue = "HeatingValue";
+
 // tslint:disable:variable-name
 
 // Energy per volume
