+++ src/units/intensity.ts
@@ -2,6 +2,8 @@ import * as UnitDivide from "../unit-divide";
 import { Watt } from "./power";
 import { SquareMeter } from "./area";
 
+export type Intensity = "Intensity";
+
 // tslint:disable:variable-name
 
 // Intensity
