+++ src/units/luminous-intensity.ts
@@ -1,0 +1,1 @@
+export type LuminousIntensity = "LuminousIntensity";
