+++ src/units/moment-of-inertia.ts
@@ -2,6 +2,8 @@ import * as UnitTimes from "../unit-times";
 import { SquareMeter } from "./area";
 import { Kilogram } from "./base-units";
 
+export type MomentOfInertia = "MomentOfInertia";
+
 // tslint:disable:variable-name
 
 /// Moment of inertia
